Move Utils package to internal/common folder
